+++ src/main/java/com/pigeonchat/lab6/services/ProfileService.java
@@ -5,6 +5,7 @@ import com.pigeonchat.lab6.dto.ProfileResponseDTO;
 import com.pigeonchat.lab6.exceptions.ProfileNotFoundException;
 import com.pigeonchat.lab6.mappers.ProfileMapper;
 import com.pigeonchat.lab6.repository.ProfileRepository;
+import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.val;
@@ -23,43 +24,45 @@ public class ProfileService {
 
     public List<ProfileResponseDTO> getAllProfiles() {
         return profileRepository.findAll().stream()
-                .map(profileMapper::toResponseDTO)
-                .collect(Collectors.toList());
+            .map(profileMapper::toResponseDTO)
+            .collect(Collectors.toList());
     }
 
-    public ProfileResponseDTO getProfileById(final UUID id) {
-        val profile = profileRepository.findById(id).orElseThrow(() -> new ProfileNotFoundException("Профиль не найден"));
+    public ProfileResponseDTO getProfileById(@NonNull final UUID id) {
+        val profile = profileRepository.findById(id)
+            .orElseThrow(() -> new ProfileNotFoundException("Профиль не найден"));
         return profileMapper.toResponseDTO(profile);
     }
 
-    public List<ProfileResponseDTO> getProfileByUsername(final String username) {
-        val profiles = profileRepository.findByUsername(username);
-        return profiles.stream()
-                .map(profileMapper::toResponseDTO)
-                .collect(Collectors.toList());
+    public List<ProfileResponseDTO> getProfileByUsername(@NonNull final String username) {
+        return profileRepository.findByUsername(username).stream()
+            .map(profileMapper::toResponseDTO)
+            .collect(Collectors.toList());
     }
 
-    public ProfileResponseDTO createProfile(final ProfileRequestDTO profileDTO) {
+    public ProfileResponseDTO createProfile(@NonNull final ProfileRequestDTO profileDTO) {
         val profile = profileMapper.toEntity(profileDTO);
-        val savedProfile = profileRepository.save(profile);
-        return profileMapper.toResponseDTO(savedProfile);
+        return profileMapper.toResponseDTO(profileRepository.save(profile));
     }
 
-    public ProfileResponseDTO updateProfile(final UUID id, final ProfileRequestDTO profileDTO) {
+    public ProfileResponseDTO updateProfile(@NonNull final UUID id, @NonNull final ProfileRequestDTO profileDTO) {
         val existingProfile = profileRepository.findById(id)
-                .orElseThrow(() -> new ProfileNotFoundException("Профиль не найден"))
-                .toBuilder()
-                .username(profileDTO.getUsername())
-                .registrationDate(profileDTO.getRegistrationDate())
-                .description(profileDTO.getDescription())
-                .avatar(profileDTO.getAvatar())
-                .email(profileDTO.getEmail())
-                .build();
-        val updatedProfile = profileRepository.save(existingProfile);
-        return profileMapper.toResponseDTO(updatedProfile);
+            .orElseThrow(() -> new ProfileNotFoundException("Профиль не найден"))
+            .toBuilder()
+            .username(profileDTO.getUsername())
+            .registrationDate(profileDTO.getRegistrationDate())
+            .description(profileDTO.getDescription())
+            .avatar(profileDTO.getAvatar())
+            .email(profileDTO.getEmail())
+            .build();
+
+        return profileMapper.toResponseDTO(profileRepository.save(existingProfile));
     }
 
-    public void deleteProfile(final UUID id) {
+    public void deleteProfile(@NonNull final UUID id) {
+        if (!profileRepository.existsById(id)) {
+            throw new ProfileNotFoundException("Профиль не найден");
+        }
         profileRepository.deleteById(id);
     }
 }
